+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Version 0.3alpha
 - rename: dl_lco -> dl_link_control_options, dl_options -> dl_lco, dl_optionsdef -> dl_options
 - added dl_link_control_options table with valid options
 - added casts for control options
+- added uri_get() and uri_set() plperlu functions for power uri handling
+- added SQL/MED standard functions DLURLSERVER() and DLURLSCHEME()
 
 Version 0.2
 -----------